feat: Allow using promises for token config
This commits allows you to do something like:
```typescript
const apiConfig : TerrasoApiConfig<{ usePromise: 'yes' }> = ...;
apiConfig.getToken().then(...);
```

What I would like to do is something like this

```typescript
type Identity<T> = t;
type WithPromise = TerrasoApiConfig<Promise>;
type WithoutPromise = TerrasoApiConfig<Identity>;
```

However, it seems this is not possible with Typescript
generics. Instead I tried using the [conditional
types](https://www.typescriptlang.org/docs/handbook/2/conditional-types.html),
which is quite messy but seems to almost work as I want it.

The current implementation seems to default to not using Promises, for
reasons that escape me (I suspect that it's because of the way
MaybePromise is implemented). This is good as it means that there
shouldn't need to be any changes to the web client. However, it means
that every time the mobile client uses `getAPIConfig`, the result
needs to be typecast to `TerrasoAPIConfig<{usePromises: 'true' }>`,
which is a real pain. In order to get around this, instead of using
the existing closure, I'm planning to use the class at the end,
initialize it at app startup, and export the functions as convenience
methods. This should let Typescript infer the types properly; I don't
think this is possible with the current closure.

I am truly not attached to any changes here, so if they seem stupid
please let me know. I'm just trying to get something working.