feat: Change getToken() type to return Promise
The mobile client needs to use Promises to fetch and store tokens,
while the web client uses a synchronous Cookie API. In order to
support both I changed the function type to return a Promise. It
created a bit of difficulty with the Redux store; the initial state
can't be set automatically because the getToken function is now
async. It will mean that the client code will have to call the
`getInitialToken` action on startup. We can maybe take another look at
this later.