feat: login changes (#20)
* feat: Allow using promises for token config

This commits allows you to do something like:
```typescript
const apiConfig : TerrasoApiConfig<{ usePromise: 'yes' }> = ...;
apiConfig.getToken().then(...);
```

What I would like to do is something like this

```typescript
type Identity<T> = t;
type WithPromise = TerrasoApiConfig<Promise>;
type WithoutPromise = TerrasoApiConfig<Identity>;
```

However, it seems this is not possible with Typescript
generics. Instead I tried using the [conditional
types](https://www.typescriptlang.org/docs/handbook/2/conditional-types.html),
which is quite messy but seems to almost work as I want it.

The current implementation seems to default to not using Promises, for
reasons that escape me (I suspect that it's because of the way
MaybePromise is implemented). This is good as it means that there
shouldn't need to be any changes to the web client. However, it means
that every time the mobile client uses `getAPIConfig`, the result
needs to be typecast to `TerrasoAPIConfig<{usePromises: 'true' }>`,
which is a real pain. In order to get around this, instead of using
the existing closure, I'm planning to use the class at the end,
initialize it at app startup, and export the functions as convenience
methods. This should let Typescript infer the types properly; I don't
think this is possible with the current closure.

I am truly not attached to any changes here, so if they seem stupid
please let me know. I'm just trying to get something working.

* Revert "feat: Allow using promises for token config"

This reverts commit 01b8e2772d495c16a528d2fd70fe99bcf9ac3b31.

* feat: Change getToken() type to return Promise

The mobile client needs to use Promises to fetch and store tokens,
while the web client uses a synchronous Cookie API. In order to
support both I changed the function type to return a Promise. It
created a bit of difficulty with the Redux store; the initial state
can't be set automatically because the getToken function is now
async. It will mean that the client code will have to call the
`getInitialToken` action on startup. We can maybe take another look at
this later.

* feat: Provide intialToken value for sync implementation

* feat: Add getToken reducer to accountSlice

* feat: Make setToken, removeToken return promises